requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-WTF>=1.2.0
WTForms>=3.1.0
Flask-Login>=0.6.0
Flask-Mail>=0.9.1
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.15.0
pyotp>=2.8.0
qrcode[pil]>=7.4.0
bcrypt>=4.0.0
email_validator>=2.0.0
psycopg2-binary>=2.9.0
gunicorn>=21.2.0
requests>=2.30.0
oauthlib>=3.2.0
